+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [commit-and-tag-version](https://github.com/absolute-version/commit-and-tag-version) for commit guidelines.
 
+## 4.8.1 (2025-03-19)
+
+
+### Bug Fixes
+
+* check ok() on StatusOr in prev_wins_ms code path
+* Disable BYOB in prod build
+
 ## 4.8.0 (2025-03-06)